+++ cli/index.js
@@ -34,6 +34,7 @@ import { printBanner } from './lib/banner.js';
 import { printHelp } from './lib/help.js';
 import { interactiveInstall, runInstallation } from './lib/installer.js';
 import { runRalph } from './lib/ralph.js';
+import { runCheck } from './lib/check.js';
 
 // Flattener module
 import { flatten as flattenCodebaseFn } from './flattener.js';
@@ -164,6 +165,11 @@ class ClaudeCraftCLI {
           // Interactive install
           await interactiveInstall(this, ctx);
         }
+        break;
+
+      case 'check':
+        printBanner(VERSION);
+        runCheck(this.config.targetPath);
         break;
 
       case 'init':